[Expr] Fixed parser bug causing a comment containing an apostrophe inside an expression block to abort parsing of the whole block. [Expr] Fixed parser bug causing `arr.<index>.<property>` to be tokenized as a decimal number, making the trailing property unreachable. [Expr] Added documentation for the `==`, `===`, `!=`, `<`, `<=`, `>` and `>=` operator aliases. [Expr] Marked `debug:context-id` as private. [Math] Fixed `math:clamp` ignoring the min and max arguments and always using the defaults. [File] Fixed `file:size` and `file:atime` returning stale values from the stat cache after a write. [Directory] Fixed `dir:remove`, `dir:remove-recursive` and `dir:rmdir` always returning null instead of the result. [Image] Fixed `image:data` not lowercasing the format when building the DATA_URI mime type. [OpenSSL] Added documentation for the `asn1:*` functions. [Gateway] Added missing documentation for the `gateway` function. [Wind] Corrected the `f` parameter charset, updated the multi-request limit and documented the `mreq` alias. [General] Added expression-level unit tests covering math, path, file, dir, stream, encoding, crypto, asn1, openssl, sys, image, gateway and the parser. [General] Added `test/run-fn.php` and updated the deployment scripts so expression tests run against the working tree instead of the installed rose-core. [General] Fixed minor bugs to increase compat with PHP 8.4. [QoL] Fixed minor issues in README, CONFIG and docstrings.

+++ src/Locale.php
@@ -87,13 +87,13 @@ class Locale
                 if (!$format) $format = 'numeric';
                 if ($config && $config->has($format)) $format = $config->get($format);
                 if ($format === 'numeric') $format = '.2,';
-                return number_format((double)$value, $format[1], $format[0], $format[2]);
+                return number_format((float)$value, $format[1], $format[0], $format[2]);
 
             case 'INTEGER':
                 if (!$format) $format = 'numeric';
                 if ($config && $config->has($format)) $format = $config->get($format);
                 if ($format === 'numeric') $format = '.2,';
-                return number_format((double)$value, 0, 0, $format[2]);
+                return number_format((float)$value, 0, 0, $format[2]);
 
             case 'TIME':
                 if (!$format) $format = 'time';
@@ -141,7 +141,7 @@ class Locale
 
                 if (Text::toUpperCase(Text::substring($formatType,0,7)) == 'NUMERIC') {
                     $tmp = $config->get($formatType);
-                    return number_format ((double)$value, $tmp[1], $tmp[0], $tmp[2]);
+                    return number_format ((float)$value, $tmp[1], $tmp[0], $tmp[2]);
                 }
 
                 return Text::format($config->get($formatType), $value);
